Ensure click event listener is always removed
The toggleCalendar() function adds/removes the event listener when called.
However, selecting a data from the calendar does not result in a call to
toggleCalendar() to close the calendar, which in turn means that the
call to removeEventListener() there is not executed.

Adding a removeEventListener() to the ngOnDestroy() call ensures that
the listener is always cleaned up.